Refactor for detailed search (#74)
* refactored for detailed search

* fixed comment

* only imports xbee library if not simulation

* fixed namespace for global variables

* forgot to import time module

* fixed imports

* fixed adds_mission function

* forgot to import mavutil
